refactor: move home from routes folder to components
